requirements.txt
pydantic
pandas
langchain
langchain-community
langchain-core
langchain-openai
python-dotenv
openai
chromadb
langchain-chroma